+++ TicTacToeTests/GameBoardTests.swift
@@ -10,29 +10,31 @@ import XCTest
 @testable import TicTacToe
 
 class GameBoardTests: XCTestCase {
+    // MARK: - Setup
+    
+    var board: GameBoard!
+    
+    override func setUp() {
+        board = GameBoard()
+    }
     
     func testCreatingEmptyBoard() {
-        let board = GameBoard()
-
-        for x in 0..<3 {
-            for y in 0..<3 {
-                XCTAssertNil(board[(x, y)])
-            }
+        loopThroughBoard { (x, y) in
+            XCTAssertNil(board[(x, y)])
         }
     }
     
     func testPlacingMarks() {
-        var board = GameBoard()
         XCTAssertNoThrow(try board.place(mark: .o, on: (0, 0)))
         XCTAssertEqual(board[(0, 0)], .o)
         XCTAssertNoThrow(try board.place(mark: .x, on: (2, 2)))
         XCTAssertEqual(board[(2, 2)], .x)
-        for x in 0..<3 {
-            for y in 0..<3 {
-                if x == 0 && y == 0 { continue }
-                if x == 2 && y == 2 { continue }
-                XCTAssertNil(board[(x, y)])
-            }
+
+        loopThroughBoard { (x, y) in
+            if (x, y) == (0, 0) { return }
+            if (x, y) == (2, 2) { return }
+            print(x, y)
+            XCTAssertNil(board[(x, y)])
         }
         
         // Make sure you can't change an already marked square
@@ -42,8 +44,6 @@ class GameBoardTests: XCTestCase {
     }
     
     func testIsFull() {
-        var board = GameBoard()
-        
         try! board.place(mark: .x, on: (0, 0))
         try! board.place(mark: .o, on: (0, 1))
         try! board.place(mark: .x, on: (1, 0))
@@ -57,5 +57,4 @@ class GameBoardTests: XCTestCase {
         try! board.place(mark: .x, on: (2, 2))
         XCTAssertTrue(board.isFull)
     }
-
 }
