+++ TicTacToeTests/GameTests.swift
@@ -1,0 +1,26 @@
+//
+//  GameTests.swift
+//  TicTacToeTests
+//
+//  Created by Jon Bash on 2019-12-11.
+//  Copyright © 2019 Lambda School. All rights reserved.
+//
+
+import XCTest
+@testable import TicTacToe
+
+class GameTests: XCTestCase {
+    var game: Game!
+
+    override func setUp() {
+        game = Game()
+    }
+
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testGame() {
+        
+    }
+}
